Patch for #348: allow longer running simulations

+++ src/FreeEnergyOutput.h
@@ -36,7 +36,7 @@ struct FreeEnergyOutput : OutputableBase {
   virtual void DoOutput(const ulong step);
 
 private:
-  void PrintData(const uint b, const uint step);
+  void PrintData(const uint b, const ulong step);
   void CalculateFreeEnergy(const uint b);
   void WriteHeader(void);
   std::string GetString(double a, uint p);
