+++ src/moves/CFCMC.h
@@ -58,7 +58,7 @@ public:
   virtual uint Prep(const double subDraw, const double movPerc);
   virtual uint Transform();
   virtual void CalcEn();
-  virtual void Accept(const uint earlyReject, const uint step);
+  virtual void Accept(const uint earlyReject, const ulong step);
   virtual void PrintAcceptKind();
 
 private:
@@ -418,7 +418,7 @@ inline double CFCMC::GetCoeff() const
 
 }
 
-inline void CFCMC::Accept(const uint rejectState, const uint step)
+inline void CFCMC::Accept(const uint rejectState, const ulong step)
 {
   GOMC_EVENT_START(1, GomcProfileEvent::ACC_CFCMC);
   bool result = false;
