+++ src/moves/Rotation.h
@@ -17,7 +17,7 @@ public:
   virtual uint Prep(const double subDraw, const double movPerc);
   virtual uint Transform();
   virtual void CalcEn();
-  virtual void Accept(const uint earlyReject, const uint step);
+  virtual void Accept(const uint earlyReject, const ulong step);
   virtual void PrintAcceptKind();
 private:
   Intermolecular inter_LJ, inter_Real, recip;
@@ -74,7 +74,7 @@ inline void Rotate::CalcEn()
   GOMC_EVENT_STOP(1, GomcProfileEvent::CALC_EN_ROTATE);
 }
 
-inline void Rotate::Accept(const uint rejectState, const uint step)
+inline void Rotate::Accept(const uint rejectState, const ulong step)
 {
   GOMC_EVENT_START(1, GomcProfileEvent::ACC_ROTATE);
   bool res = false;
