+++ src/moves/TargetedSwap.h
@@ -202,7 +202,7 @@ public:
   virtual uint Prep(const double subDraw, const double movPerc);
   virtual uint Transform();
   virtual void CalcEn();
-  virtual void Accept(const uint earlyReject, const uint step);
+  virtual void Accept(const uint earlyReject, const ulong step);
   virtual void PrintAcceptKind();
 
 private:
@@ -637,7 +637,7 @@ inline double TargetedSwap::GetCoeff() const
 #endif
 }
 
-inline void TargetedSwap::Accept(const uint rejectState, const uint step)
+inline void TargetedSwap::Accept(const uint rejectState, const ulong step)
 {
   bool result;
   //If we didn't skip the move calculation
